put myTrips array into separate file, changed date formats
