+++ tripscribe/src/App.js
@@ -1,4 +1,5 @@
 import './App.css';
+import MyTrips from './pages/myTrips/myTrips.jsx';
 
 function App() {
   return (
@@ -38,7 +39,9 @@ function App() {
               <p>My favourite trip was when I backpacked South East Asia solo after uni. I did Thailand, Vietnam, Cambodia, Laos, Bali and the Philippines. It took 5 months and it was so much fun, I met so many amazing people. My favourite country was Vietnam!</p>
             </li>
         </ul>
-  
+      <div>
+        <MyTrips />
+      </div>
       </div>
     );
   }
