+++ tripscribe/src/pages/myTrips/tripsArray.js
@@ -1,0 +1,63 @@
+import NewYork from "./images/newyork.jpg"
+import Paris from "./images/paris.jpg"
+import Lisbon from "./images/Lisbon.jpg"
+import Berlin from "./images/Berlin.jpg"
+
+const tripsArray = [
+	{
+	  image: Paris,
+	  city: 'Paris',
+	  country: 'France',
+	  title: 'Trip to Paris',
+	  startDate: '2024-06-01',
+	  endDate: '2024-06-10',
+	  description: 'A wonderful trip to Paris with visits to the Eiffel Tower and the Louvre.'
+	},
+	{
+	  image: NewYork,
+	  city: 'New York',
+	  country: 'USA',
+	  title: 'Trip to New York',
+	  startDate: '2024-07-15',
+	  endDate: '2024-07-20',
+	  description: 'Exploring the city that never sleeps, including Times Square and Central Park.'
+	},
+	{
+		image: Lisbon,
+		city: 'Lisbon',
+		country: 'Spain',
+		title: 'Trip to Lisbon',
+		startDate: '2024-07-15',
+		endDate: '2024-07-20',
+		description: 'Exploring the city that never sleeps, including Times Square and Central Park.'
+	},
+	{
+		image: Lisbon,
+		city: 'Lisbon',
+		country: 'Spain',
+		title: 'Trip to Lisbon',
+		startDate: '2024-07-15',
+		endDate: '2024-07-20',
+		description: 'Exploring the city that never sleeps, including Times Square and Central Park.'
+	},
+	{
+		image: NewYork,
+		city: 'New York',
+		country: 'USA',
+		title: 'Trip to New York',
+		startDate: '2024-05-15',
+		endDate: '2024-05-20',
+		description: 'Exploring the city that never sleeps, including Times Square and Central Park.'
+	  },
+	  {
+		image: Berlin,
+		city: 'Berlin',
+		country: 'Germany',
+		title: 'Trip to Lisbon',
+		startDate: '2024-07-15',
+		endDate: '2024-07-20',
+		description: 'Exploring the city that never sleeps, including Times Square and Central Park.'
+	}
+];
+
+export default tripsArray;
